% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgraph.R
\name{lightgraphOutput}
\alias{lightgraphOutput}
\title{Shiny output binding for lightgraph}
\usage{
lightgraphOutput(outputId, width = "100\%", height = "800px")
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width}{Width of the visualization (must be a valid CSS unit)}

\item{height}{Height of the visualization (must be a valid CSS unit)}
}
\value{
An output element for use in Shiny UI
}
\description{
Output function for rendering lightgraph visualizations in Shiny applications.
}
