% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgraph.R
\name{renderLightgraph}
\alias{renderLightgraph}
\title{Shiny render function for lightgraph}
\usage{
renderLightgraph(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a lightgraph visualization}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression (with quote())?}
}
\value{
A render function for use in Shiny server
}
\description{
Render function for creating lightgraph visualizations in Shiny applications.
}
